fix: keep installment share allocation typed

+++ app/lib/features/transactions/transaction/usecases/create_transaction_usecase.dart
@@ -333,7 +333,8 @@ class CreateTransactionUseCase {
                             installmentInCents /
                             totalInCents)
                         .round()
-                        .clamp(0, remaining);
+                        .clamp(0, remaining)
+                        .toInt();
 
         shares[memberId] = memberCents / 100;
         remainingByMember[memberId] =
